Circle Collisions supported, collision betweens more complex polygons supported
Note: To create more complex gemoetry, currently you must create a new shapes class that extends my shape class in the chowshapes package. I will simplyify this process to allow for custom size polygons with any number of vertices.

Modified the scope of most class concstructors to public. Not sure why they weren't to begin with but now the program can be exported to a JAR file.

Included a Pair algoirthm for checking two different actors for collision to avoid duplicate computations.

Changed rendering order. Physics is first calculated before the game window is rendered.

Included hasCollidedCircle for two polygons that are of the shape of circles. -Fully working.

hasCollidedPolygonAndCircle is still in development. There is a lot of weird behavior that I cannot quite nail down right now. I am currently looking for outside help on this issue.

Triangles and Hexagons are now supported in the engine.

Circles are only supported if the other objects are circles.